Move the util package, refactor of waitUntil
