+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/combat/AutoArmor.kt
@@ -13,6 +13,7 @@ import net.ccbluex.liquidbounce.features.module.Category
 import net.ccbluex.liquidbounce.features.module.Module
 import net.ccbluex.liquidbounce.features.module.modules.player.InventoryCleaner.canBeRepairedWithOther
 import net.ccbluex.liquidbounce.utils.client.PacketUtils.sendPacket
+import net.ccbluex.liquidbounce.utils.kotlin.waitUntil
 import net.ccbluex.liquidbounce.utils.inventory.ArmorComparator.getBestArmorSet
 import net.ccbluex.liquidbounce.utils.inventory.InventoryManager
 import net.ccbluex.liquidbounce.utils.inventory.InventoryManager.autoArmorCurrentSlot
@@ -25,7 +26,6 @@ import net.ccbluex.liquidbounce.utils.inventory.InventoryUtils.serverOpenInvento
 import net.ccbluex.liquidbounce.utils.inventory.InventoryUtils.toHotbarIndex
 import net.ccbluex.liquidbounce.utils.inventory.SilentHotbar
 import net.ccbluex.liquidbounce.utils.inventory.hasItemAgePassed
-import net.ccbluex.liquidbounce.utils.kotlin.CoroutineUtils.waitUntil
 import net.ccbluex.liquidbounce.utils.timing.TimeUtils.randomDelay
 import net.minecraft.client.gui.inventory.GuiInventory
 import net.minecraft.entity.EntityLiving.getArmorPosition
@@ -146,7 +146,7 @@ object AutoArmor : Module("AutoArmor", Category.COMBAT, hideModule = false) {
         // Not really needed to bypass
         delay(randomDelay(minDelay, maxDelay).toLong())
 
-        waitUntil(TickScheduler::isEmpty)
+        waitUntil { TickScheduler.isEmpty() }
 
         // Sync selected slot next tick
         if (hasClickedHotbar)
@@ -234,7 +234,7 @@ object AutoArmor : Module("AutoArmor", Category.COMBAT, hideModule = false) {
         }
 
         // Wait till all scheduled clicks were sent
-        waitUntil(TickScheduler::isEmpty)
+        waitUntil { TickScheduler.isEmpty() }
     }
 
     fun equipFromHotbarInChest(hotbarIndex: Int?, stack: ItemStack) {
