+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/ServerCrasher.kt
@@ -16,6 +16,7 @@ import net.ccbluex.liquidbounce.utils.client.PacketUtils.sendPacket
 import net.ccbluex.liquidbounce.utils.extensions.*
 import net.ccbluex.liquidbounce.utils.kotlin.RandomUtils.randomNumber
 import net.ccbluex.liquidbounce.utils.kotlin.RandomUtils.randomString
+import net.ccbluex.liquidbounce.utils.kotlin.SharedScopes
 import net.ccbluex.liquidbounce.utils.timing.MSTimer
 import net.minecraft.init.Items
 import net.minecraft.item.ItemStack
