+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/misc/StaffDetector.kt
@@ -19,7 +19,7 @@ import net.ccbluex.liquidbounce.features.module.Category
 import net.ccbluex.liquidbounce.features.module.Module
 import net.ccbluex.liquidbounce.ui.client.hud.element.elements.Notification
 import net.ccbluex.liquidbounce.utils.client.chat
-import net.ccbluex.liquidbounce.utils.extensions.SharedScopes
+import net.ccbluex.liquidbounce.utils.kotlin.SharedScopes
 import net.ccbluex.liquidbounce.utils.io.HttpUtils
 import net.minecraft.entity.Entity
 import net.minecraft.entity.player.EntityPlayer
